Fix sequence of mappings parsing
When a sequence item (after '-') is followed by a newline and indented
content, the parser now correctly parses the indented content as the
sequence item value. This fixes parsing of structures like:

-
  name: value
  key: value

The logic was added to parse_sequence_with_base_indent to mirror the
mapping value parsing that handles indented values.